requirements.txt
streamlit
pandas
google-genai
python-docx
fpdf
supabase
